#10 [MOVE] GroupCreate 및 GroupCreateSuccess 패키지 위치 변경
