+++ presentation/src/main/java/com/sopt/presentation/groupCreate/groupCreateSuccess/GroupCreateSuccessSideEffect.kt
@@ -1,4 +1,4 @@
-package com.sopt.presentation.group.groupCreateSuccess
+package com.sopt.presentation.groupCreate.groupCreateSuccess
 
 sealed interface GroupCreateSuccessSideEffect {
     data class NavigateToGroupDetail(val groupId: Long) : GroupCreateSuccessSideEffect
