+++ presentation/src/main/java/com/sopt/presentation/groupCreate/groupCreateSuccess/navigation/GroupCreateSuccessNavigation.kt
@@ -1,12 +1,12 @@
-package com.sopt.presentation.group.groupCreateSuccess.navigation
+package com.sopt.presentation.groupCreate.groupCreateSuccess.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.sopt.core.navigation.Route
-import com.sopt.presentation.group.groupCreate.navigation.GroupCreate
-import com.sopt.presentation.group.groupCreateSuccess.GroupCreateSuccessRoute
+import com.sopt.presentation.groupCreate.navigation.GroupCreate
+import com.sopt.presentation.groupCreate.groupCreateSuccess.GroupCreateSuccessRoute
 import com.sopt.presentation.groupDetail.navigation.navigateGroupDetail
 import kotlinx.serialization.Serializable
