+++ presentation/src/main/java/com/sopt/presentation/groupCreate/navigation/GroupCreateNavigation.kt
@@ -1,4 +1,4 @@
-package com.sopt.presentation.group.groupCreate.navigation
+package com.sopt.presentation.groupCreate.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavController
@@ -6,8 +6,8 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.sopt.core.navigation.Route
-import com.sopt.presentation.group.groupCreate.GroupCreateRoute
-import com.sopt.presentation.group.groupCreateSuccess.navigation.navigateToGroupCreateSuccess
+import com.sopt.presentation.groupCreate.GroupCreateRoute
+import com.sopt.presentation.groupCreate.groupCreateSuccess.navigation.navigateToGroupCreateSuccess
 import kotlinx.serialization.Serializable
 
 fun NavController.navigateToGroupCreate(navOptions: NavOptions? = null) {
