+++ presentation/src/main/java/com/sopt/presentation/groupCreate/permission/rememberGalleryLauncher.kt
@@ -1,4 +1,4 @@
-package com.sopt.presentation.group.groupCreate.permission
+package com.sopt.presentation.groupCreate.permission
 
 import android.net.Uri
 import androidx.activity.compose.ManagedActivityResultLauncher
