+++ presentation/src/main/java/com/sopt/presentation/main/MainScreen.kt
@@ -40,8 +40,8 @@ import com.sopt.core.designsystem.component.snackbar.NoostakSnackBar
 import com.sopt.core.util.NoRippleInteractionSource
 import com.sopt.presentation.calendar.navigation.calendarNavGraph
 import com.sopt.presentation.example.navigation.exampleNavGraph
-import com.sopt.presentation.group.groupCreate.navigation.groupCreateNavGraph
-import com.sopt.presentation.group.groupCreateSuccess.navigation.groupCreateSuccessNavGraph
+import com.sopt.presentation.groupCreate.navigation.groupCreateNavGraph
+import com.sopt.presentation.groupCreate.groupCreateSuccess.navigation.groupCreateSuccessNavGraph
 import com.sopt.presentation.group.navigation.groupNavGraph
 import com.sopt.presentation.groupDetail.navigation.groupDetailNavGraph
 import com.sopt.presentation.mypage.navigation.myPageNavGraph
